+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.6.3'
+__version__ = '0.6.4'
 
 
 __submodules__ = {